requirements.txt
# Core web framework
flask>=2.0.0

# Computer vision and video processing
opencv-python>=4.5.0
ultralytics>=8.0.0

# Audio processing
ffmpeg-python>=0.2.0

# Speech recognition
openai-whisper>=20240930

# Translation service
deep-translator>=1.9.0

# Additional dependencies that may be needed
pathlib2>=2.3.0
numpy>=1.21.0
Pillow>=8.0.0